backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Scraping
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0

# LLM
openai==1.10.0

# PDF Generation
weasyprint==60.2
jinja2==3.1.3

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Payments
stripe==7.8.0

# JWT Authentication
pyjwt==2.8.0

